Add files via upload
This repository contains the code for a simple and effective landing page. It's a result of my hands-on learning and practical experience in web development during my internship. 💼

Feel free to explore the code, provide feedback, or even use it as a reference for your own projects. Collaboration and growth are at the heart of open-source development! 🌐